Modificado de metodo get custom_user_social_media
se modifico el metodo get de  custom_user_social_media para que retorne la img dependiendo del modelo type

+++ profile/services/customer_user_socialmedia_service.py
@@ -50,7 +50,7 @@ class SocialMediaService:
     
         # busca el objeto y si lo encuentra lo borra 
         cusm = get_object_or_404(CustomerUserCustomSocialMedia, id=pk, customer_user_id=customer_user)
-        if cusm.image != "custom_social_media/undefined.png":
+        if cusm.type=="socialmedia" and cusm.image != "custom_social_media/undefined.png":
             try:
                 os.remove(cusm.image.path)
             except Exception as e:
